feat(telemetry): add rindo config to telemetry object
this adds an anonymized copy of the Rindo configuration to the
telemetry object, so that various configuration options can be tracked
and observed over time (while preserving anonymity). this is being done
in preparation for Rindo v3, so that we may get better to insights as
to which flags are being used (i.e. are they being used or not) in
order to determine what default flags we wish to set in v3

Add Rindo Configuration Contents to Telemetry